web/requirements.txt
fastapi==0.115.12
jinja2==3.1.6
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.0
